Entity branch polish before merging (#651)
* Clarified bit of parsing code

* Remove extra 'ensureFormDef'

* Merged all dataset/entity migrations into one file

* Move a form dataset attachment resource into forms.js

* Deleted uneeded query module tets

* Add outer describe block around dataset tests

* Reorganize tests, including moving forms/datasets.js tests to other file

* Reorged entity worker tests
