+++ lib/data/entity.js
@@ -57,7 +57,7 @@ const parseSubmissionXml = (entityFields, xml) => new Promise((resolve, reject) 
   const stream = submissionXmlToFieldStream(entityFields, xml, true);
   stream.on('error', reject);
   stream.on('data', ({ field, text }) => {
-    if (field.name === 'entity' && field.attrs !== {}) {
+    if (field.name === 'entity') {
       entity.system.dataset = (field.attrs.dataset ? field.attrs.dataset : field.attrs['entities:dataset']);
     } else if (field.path.indexOf('/meta/entity') === 0)
       entity.system[field.name] = text;
