+++ lib/data/submission.js
@@ -14,7 +14,7 @@ const fmdiff = require('fast-myers-diff');
 const { SchemaStack } = require('./schema');
 const { noop } = require('../util/util');
 const { stripNamespacesFromPath } = require('../util/xml');
-const { union, last, pluck, isEmpty } = require('ramda');
+const { union, last, pluck } = require('ramda');
 
 
 // reads submission xml with the streaming parser, and outputs a stream of every
@@ -37,7 +37,10 @@ const submissionXmlToFieldStream = (fields, xml, includeStructuralAttrs = false)
       const field = stack.push(name);
       if (field != null) {
         textBuffer = '';
-        if (includeStructuralAttrs && attrs !== {} && !isEmpty(field) && field.isStructural())
+        // If the field is a structural field AND it has attributes AND we should output them, THEN do so.
+        if (includeStructuralAttrs &&
+          (typeof field.isStructural === 'function' && field.isStructural()) &&
+          Object.keys(attrs).length !== 0)
           outStream.push({ field: { ...field, attrs }, text: null });
       }
     },
