+++ lib/model/query/forms.js
@@ -64,7 +64,6 @@ const createNew = (partial, project, publish = false) => ({ run, Datasets, FormA
     getFormFields(partial.xml),
     getDataset(partial.xml)
   ])
-    // eslint-disable-next-line no-unused-vars
     .then(([ keyId, fields, dataset ]) => Forms._createNew(partial, partial.def.with({ keyId }), project, publish)
       .then((savedForm) => {
         const ids = { formId: savedForm.id, formDefId: savedForm.def.id };
