+++ lib/resources/datasets.js
@@ -8,7 +8,6 @@
 // except according to the terms contained in the LICENSE file.
 
 const sanitize = require('sanitize-filename');
-const { Form } = require('../model/frames');
 const { getOrNotFound } = require('../util/promise');
 const { streamEntityCsvs } = require('../data/entity');
 const { contentDisposition } = require('../util/http');
@@ -34,14 +33,4 @@ module.exports = (service, endpoint) => {
             response.append('Content-Type', 'text/csv');
             return streamEntityCsvs(entities, dataset.properties);
           })))));
-
-  service.patch('/projects/:projectId/forms/:id/draft/attachments/:name', endpoint(({ Datasets, FormAttachments, Forms }, { auth, params, body }) =>
-    Forms.getByProjectAndXmlFormId(params.projectId, params.id, false, Form.DraftVersion)
-      .then(getOrNotFound)
-      .then((form) => auth.canOrReject('form.update', form))
-      .then((form) => Promise.all([
-        Datasets.getByProjectAndName(params.projectId, params.name.replace(/\.csv$/, '')).then(getOrNotFound),
-        FormAttachments.getByFormDefIdAndName(form.draftDefId, params.name).then(getOrNotFound)
-      ])
-        .then(([dataset, attachment]) => FormAttachments.update(form, attachment, null, body.dataset ? dataset.id : null)))));
 };
