+++ lib/resources/forms.js
@@ -184,6 +184,17 @@ module.exports = (service, endpoint) => {
       .then(([form, ]) => ensureDef(form))
       .then(() => Datasets.getDatasetDiff(params.projectId, params.id, false))));
 
+  service.patch('/projects/:projectId/forms/:id/draft/attachments/:name', endpoint(({ Datasets, FormAttachments, Forms }, { auth, params, body }) =>
+    Forms.getByProjectAndXmlFormId(params.projectId, params.id, false, Form.DraftVersion)
+      .then(getOrNotFound)
+      .then((form) => auth.canOrReject('form.update', form))
+      .then((form) => Promise.all([
+        Datasets.getByProjectAndName(params.projectId, params.name.replace(/\.csv$/, '')).then(getOrNotFound),
+        FormAttachments.getByFormDefIdAndName(form.draftDefId, params.name).then(getOrNotFound)
+      ])
+        .then(([dataset, attachment]) => FormAttachments.update(form, attachment, null, body.dataset ? dataset.id : null)))));
+
+
   service.delete('/projects/:projectId/forms/:id/draft', endpoint(({ Audits, Forms, Submissions }, { params, auth }) =>
     Forms.getByProjectAndXmlFormId(params.projectId, params.id)
       .then(getOrNotFound)
