+++ lib/util/db.js
@@ -11,11 +11,11 @@ const { inspect } = require('util');
 const { mergeRight, pick, always } = require('ramda');
 const { sql } = require('slonik');
 const { raw } = require('slonik-sql-tag-raw');
-const { reject, rejectIf } = require('./promise');
+const { reject } = require('./promise');
 const Problem = require('./problem');
 const Option = require('./option');
 const { PartialPipe, mapStream } = require('./stream');
-const { construct, noargs } = require('./util');
+const { construct } = require('./util');
 const { isTrue, isFalse } = require('./http');
 
 const { Transform } = require('stream');
@@ -497,14 +497,11 @@ const postgresErrorToProblem = (x) => {
   return reject(error);
 };
 
-// To ensure the form we fetched successfully joined to the intended def.
-const ensureFormDef = rejectIf(((form) => form.def.id == null), noargs(Problem.user.notFound));
-
 module.exports = {
   connectionString, connectionObject,
   unjoiner, extender, equals, page, queryFuncs,
   insert, insertMany, updater, markDeleted, markUndeleted,
-  QueryOptions, ensureFormDef,
+  QueryOptions,
   postgresErrorToProblem
 };
